tester/requirements.txt
pyyaml
requests
jsonschema
